Generating the Full Design Matrix

getFullDesign is used to generate the design matrix for linear regression with ABCs.

Usage

getFullDesign(formula, data, center = TRUE)

Arguments

formula

an object of class "formula()" (or one that can be coerced to that class); a symbolic description of the model to be fitted.

data

a data frame (or object coercible by as.data.frame to a data frame) containing the variables in the model.

center

Boolean, whether to center continuous predictors. TRUE by default for lmabc.

Details

The full design matrix is similar to the design matrix in standard linear regression. However, instead of removing one level from each categorical variable, getFullDesign retains all levels by manipulating contrast matrices.

The full design matrix is used with the constraint matrix from getConstraints() to fit a model with the ABCs.

This method is called by lmabc(). This method is useful for implementing additional variations of the ABCs.

Value

getFullDesign returns a matrix representing the full design matrix for a particular regression model.

See also

Examples

# full design matrix
getFullDesign(Sepal.Length ~ Petal.Length + Species + Petal.Length*Species, data = iris)
#>     (Intercept) Petal.Length Speciessetosa Speciesversicolor Speciesvirginica
#> 1             1       -2.358             1                 0                0
#> 2             1       -2.358             1                 0                0
#> 3             1       -2.458             1                 0                0
#> 4             1       -2.258             1                 0                0
#> 5             1       -2.358             1                 0                0
#> 6             1       -2.058             1                 0                0
#> 7             1       -2.358             1                 0                0
#> 8             1       -2.258             1                 0                0
#> 9             1       -2.358             1                 0                0
#> 10            1       -2.258             1                 0                0
#> 11            1       -2.258             1                 0                0
#> 12            1       -2.158             1                 0                0
#> 13            1       -2.358             1                 0                0
#> 14            1       -2.658             1                 0                0
#> 15            1       -2.558             1                 0                0
#> 16            1       -2.258             1                 0                0
#> 17            1       -2.458             1                 0                0
#> 18            1       -2.358             1                 0                0
#> 19            1       -2.058             1                 0                0
#> 20            1       -2.258             1                 0                0
#> 21            1       -2.058             1                 0                0
#> 22            1       -2.258             1                 0                0
#> 23            1       -2.758             1                 0                0
#> 24            1       -2.058             1                 0                0
#> 25            1       -1.858             1                 0                0
#> 26            1       -2.158             1                 0                0
#> 27            1       -2.158             1                 0                0
#> 28            1       -2.258             1                 0                0
#> 29            1       -2.358             1                 0                0
#> 30            1       -2.158             1                 0                0
#> 31            1       -2.158             1                 0                0
#> 32            1       -2.258             1                 0                0
#> 33            1       -2.258             1                 0                0
#> 34            1       -2.358             1                 0                0
#> 35            1       -2.258             1                 0                0
#> 36            1       -2.558             1                 0                0
#> 37            1       -2.458             1                 0                0
#> 38            1       -2.358             1                 0                0
#> 39            1       -2.458             1                 0                0
#> 40            1       -2.258             1                 0                0
#> 41            1       -2.458             1                 0                0
#> 42            1       -2.458             1                 0                0
#> 43            1       -2.458             1                 0                0
#> 44            1       -2.158             1                 0                0
#> 45            1       -1.858             1                 0                0
#> 46            1       -2.358             1                 0                0
#> 47            1       -2.158             1                 0                0
#> 48            1       -2.358             1                 0                0
#> 49            1       -2.258             1                 0                0
#> 50            1       -2.358             1                 0                0
#> 51            1        0.942             0                 1                0
#> 52            1        0.742             0                 1                0
#> 53            1        1.142             0                 1                0
#> 54            1        0.242             0                 1                0
#> 55            1        0.842             0                 1                0
#> 56            1        0.742             0                 1                0
#> 57            1        0.942             0                 1                0
#> 58            1       -0.458             0                 1                0
#> 59            1        0.842             0                 1                0
#> 60            1        0.142             0                 1                0
#> 61            1       -0.258             0                 1                0
#> 62            1        0.442             0                 1                0
#> 63            1        0.242             0                 1                0
#> 64            1        0.942             0                 1                0
#> 65            1       -0.158             0                 1                0
#> 66            1        0.642             0                 1                0
#> 67            1        0.742             0                 1                0
#> 68            1        0.342             0                 1                0
#> 69            1        0.742             0                 1                0
#> 70            1        0.142             0                 1                0
#> 71            1        1.042             0                 1                0
#> 72            1        0.242             0                 1                0
#> 73            1        1.142             0                 1                0
#> 74            1        0.942             0                 1                0
#> 75            1        0.542             0                 1                0
#> 76            1        0.642             0                 1                0
#> 77            1        1.042             0                 1                0
#> 78            1        1.242             0                 1                0
#> 79            1        0.742             0                 1                0
#> 80            1       -0.258             0                 1                0
#> 81            1        0.042             0                 1                0
#> 82            1       -0.058             0                 1                0
#> 83            1        0.142             0                 1                0
#> 84            1        1.342             0                 1                0
#> 85            1        0.742             0                 1                0
#> 86            1        0.742             0                 1                0
#> 87            1        0.942             0                 1                0
#> 88            1        0.642             0                 1                0
#> 89            1        0.342             0                 1                0
#> 90            1        0.242             0                 1                0
#> 91            1        0.642             0                 1                0
#> 92            1        0.842             0                 1                0
#> 93            1        0.242             0                 1                0
#> 94            1       -0.458             0                 1                0
#> 95            1        0.442             0                 1                0
#> 96            1        0.442             0                 1                0
#> 97            1        0.442             0                 1                0
#> 98            1        0.542             0                 1                0
#> 99            1       -0.758             0                 1                0
#> 100           1        0.342             0                 1                0
#> 101           1        2.242             0                 0                1
#> 102           1        1.342             0                 0                1
#> 103           1        2.142             0                 0                1
#> 104           1        1.842             0                 0                1
#> 105           1        2.042             0                 0                1
#> 106           1        2.842             0                 0                1
#> 107           1        0.742             0                 0                1
#> 108           1        2.542             0                 0                1
#> 109           1        2.042             0                 0                1
#> 110           1        2.342             0                 0                1
#> 111           1        1.342             0                 0                1
#> 112           1        1.542             0                 0                1
#> 113           1        1.742             0                 0                1
#> 114           1        1.242             0                 0                1
#> 115           1        1.342             0                 0                1
#> 116           1        1.542             0                 0                1
#> 117           1        1.742             0                 0                1
#> 118           1        2.942             0                 0                1
#> 119           1        3.142             0                 0                1
#> 120           1        1.242             0                 0                1
#> 121           1        1.942             0                 0                1
#> 122           1        1.142             0                 0                1
#> 123           1        2.942             0                 0                1
#> 124           1        1.142             0                 0                1
#> 125           1        1.942             0                 0                1
#> 126           1        2.242             0                 0                1
#> 127           1        1.042             0                 0                1
#> 128           1        1.142             0                 0                1
#> 129           1        1.842             0                 0                1
#> 130           1        2.042             0                 0                1
#> 131           1        2.342             0                 0                1
#> 132           1        2.642             0                 0                1
#> 133           1        1.842             0                 0                1
#> 134           1        1.342             0                 0                1
#> 135           1        1.842             0                 0                1
#> 136           1        2.342             0                 0                1
#> 137           1        1.842             0                 0                1
#> 138           1        1.742             0                 0                1
#> 139           1        1.042             0                 0                1
#> 140           1        1.642             0                 0                1
#> 141           1        1.842             0                 0                1
#> 142           1        1.342             0                 0                1
#> 143           1        1.342             0                 0                1
#> 144           1        2.142             0                 0                1
#> 145           1        1.942             0                 0                1
#> 146           1        1.442             0                 0                1
#> 147           1        1.242             0                 0                1
#> 148           1        1.442             0                 0                1
#> 149           1        1.642             0                 0                1
#> 150           1        1.342             0                 0                1
#>     Petal.Length:Speciessetosa Petal.Length:Speciesversicolor
#> 1                       -2.358                          0.000
#> 2                       -2.358                          0.000
#> 3                       -2.458                          0.000
#> 4                       -2.258                          0.000
#> 5                       -2.358                          0.000
#> 6                       -2.058                          0.000
#> 7                       -2.358                          0.000
#> 8                       -2.258                          0.000
#> 9                       -2.358                          0.000
#> 10                      -2.258                          0.000
#> 11                      -2.258                          0.000
#> 12                      -2.158                          0.000
#> 13                      -2.358                          0.000
#> 14                      -2.658                          0.000
#> 15                      -2.558                          0.000
#> 16                      -2.258                          0.000
#> 17                      -2.458                          0.000
#> 18                      -2.358                          0.000
#> 19                      -2.058                          0.000
#> 20                      -2.258                          0.000
#> 21                      -2.058                          0.000
#> 22                      -2.258                          0.000
#> 23                      -2.758                          0.000
#> 24                      -2.058                          0.000
#> 25                      -1.858                          0.000
#> 26                      -2.158                          0.000
#> 27                      -2.158                          0.000
#> 28                      -2.258                          0.000
#> 29                      -2.358                          0.000
#> 30                      -2.158                          0.000
#> 31                      -2.158                          0.000
#> 32                      -2.258                          0.000
#> 33                      -2.258                          0.000
#> 34                      -2.358                          0.000
#> 35                      -2.258                          0.000
#> 36                      -2.558                          0.000
#> 37                      -2.458                          0.000
#> 38                      -2.358                          0.000
#> 39                      -2.458                          0.000
#> 40                      -2.258                          0.000
#> 41                      -2.458                          0.000
#> 42                      -2.458                          0.000
#> 43                      -2.458                          0.000
#> 44                      -2.158                          0.000
#> 45                      -1.858                          0.000
#> 46                      -2.358                          0.000
#> 47                      -2.158                          0.000
#> 48                      -2.358                          0.000
#> 49                      -2.258                          0.000
#> 50                      -2.358                          0.000
#> 51                       0.000                          0.942
#> 52                       0.000                          0.742
#> 53                       0.000                          1.142
#> 54                       0.000                          0.242
#> 55                       0.000                          0.842
#> 56                       0.000                          0.742
#> 57                       0.000                          0.942
#> 58                       0.000                         -0.458
#> 59                       0.000                          0.842
#> 60                       0.000                          0.142
#> 61                       0.000                         -0.258
#> 62                       0.000                          0.442
#> 63                       0.000                          0.242
#> 64                       0.000                          0.942
#> 65                       0.000                         -0.158
#> 66                       0.000                          0.642
#> 67                       0.000                          0.742
#> 68                       0.000                          0.342
#> 69                       0.000                          0.742
#> 70                       0.000                          0.142
#> 71                       0.000                          1.042
#> 72                       0.000                          0.242
#> 73                       0.000                          1.142
#> 74                       0.000                          0.942
#> 75                       0.000                          0.542
#> 76                       0.000                          0.642
#> 77                       0.000                          1.042
#> 78                       0.000                          1.242
#> 79                       0.000                          0.742
#> 80                       0.000                         -0.258
#> 81                       0.000                          0.042
#> 82                       0.000                         -0.058
#> 83                       0.000                          0.142
#> 84                       0.000                          1.342
#> 85                       0.000                          0.742
#> 86                       0.000                          0.742
#> 87                       0.000                          0.942
#> 88                       0.000                          0.642
#> 89                       0.000                          0.342
#> 90                       0.000                          0.242
#> 91                       0.000                          0.642
#> 92                       0.000                          0.842
#> 93                       0.000                          0.242
#> 94                       0.000                         -0.458
#> 95                       0.000                          0.442
#> 96                       0.000                          0.442
#> 97                       0.000                          0.442
#> 98                       0.000                          0.542
#> 99                       0.000                         -0.758
#> 100                      0.000                          0.342
#> 101                      0.000                          0.000
#> 102                      0.000                          0.000
#> 103                      0.000                          0.000
#> 104                      0.000                          0.000
#> 105                      0.000                          0.000
#> 106                      0.000                          0.000
#> 107                      0.000                          0.000
#> 108                      0.000                          0.000
#> 109                      0.000                          0.000
#> 110                      0.000                          0.000
#> 111                      0.000                          0.000
#> 112                      0.000                          0.000
#> 113                      0.000                          0.000
#> 114                      0.000                          0.000
#> 115                      0.000                          0.000
#> 116                      0.000                          0.000
#> 117                      0.000                          0.000
#> 118                      0.000                          0.000
#> 119                      0.000                          0.000
#> 120                      0.000                          0.000
#> 121                      0.000                          0.000
#> 122                      0.000                          0.000
#> 123                      0.000                          0.000
#> 124                      0.000                          0.000
#> 125                      0.000                          0.000
#> 126                      0.000                          0.000
#> 127                      0.000                          0.000
#> 128                      0.000                          0.000
#> 129                      0.000                          0.000
#> 130                      0.000                          0.000
#> 131                      0.000                          0.000
#> 132                      0.000                          0.000
#> 133                      0.000                          0.000
#> 134                      0.000                          0.000
#> 135                      0.000                          0.000
#> 136                      0.000                          0.000
#> 137                      0.000                          0.000
#> 138                      0.000                          0.000
#> 139                      0.000                          0.000
#> 140                      0.000                          0.000
#> 141                      0.000                          0.000
#> 142                      0.000                          0.000
#> 143                      0.000                          0.000
#> 144                      0.000                          0.000
#> 145                      0.000                          0.000
#> 146                      0.000                          0.000
#> 147                      0.000                          0.000
#> 148                      0.000                          0.000
#> 149                      0.000                          0.000
#> 150                      0.000                          0.000
#>     Petal.Length:Speciesvirginica
#> 1                           0.000
#> 2                           0.000
#> 3                           0.000
#> 4                           0.000
#> 5                           0.000
#> 6                           0.000
#> 7                           0.000
#> 8                           0.000
#> 9                           0.000
#> 10                          0.000
#> 11                          0.000
#> 12                          0.000
#> 13                          0.000
#> 14                          0.000
#> 15                          0.000
#> 16                          0.000
#> 17                          0.000
#> 18                          0.000
#> 19                          0.000
#> 20                          0.000
#> 21                          0.000
#> 22                          0.000
#> 23                          0.000
#> 24                          0.000
#> 25                          0.000
#> 26                          0.000
#> 27                          0.000
#> 28                          0.000
#> 29                          0.000
#> 30                          0.000
#> 31                          0.000
#> 32                          0.000
#> 33                          0.000
#> 34                          0.000
#> 35                          0.000
#> 36                          0.000
#> 37                          0.000
#> 38                          0.000
#> 39                          0.000
#> 40                          0.000
#> 41                          0.000
#> 42                          0.000
#> 43                          0.000
#> 44                          0.000
#> 45                          0.000
#> 46                          0.000
#> 47                          0.000
#> 48                          0.000
#> 49                          0.000
#> 50                          0.000
#> 51                          0.000
#> 52                          0.000
#> 53                          0.000
#> 54                          0.000
#> 55                          0.000
#> 56                          0.000
#> 57                          0.000
#> 58                          0.000
#> 59                          0.000
#> 60                          0.000
#> 61                          0.000
#> 62                          0.000
#> 63                          0.000
#> 64                          0.000
#> 65                          0.000
#> 66                          0.000
#> 67                          0.000
#> 68                          0.000
#> 69                          0.000
#> 70                          0.000
#> 71                          0.000
#> 72                          0.000
#> 73                          0.000
#> 74                          0.000
#> 75                          0.000
#> 76                          0.000
#> 77                          0.000
#> 78                          0.000
#> 79                          0.000
#> 80                          0.000
#> 81                          0.000
#> 82                          0.000
#> 83                          0.000
#> 84                          0.000
#> 85                          0.000
#> 86                          0.000
#> 87                          0.000
#> 88                          0.000
#> 89                          0.000
#> 90                          0.000
#> 91                          0.000
#> 92                          0.000
#> 93                          0.000
#> 94                          0.000
#> 95                          0.000
#> 96                          0.000
#> 97                          0.000
#> 98                          0.000
#> 99                          0.000
#> 100                         0.000
#> 101                         2.242
#> 102                         1.342
#> 103                         2.142
#> 104                         1.842
#> 105                         2.042
#> 106                         2.842
#> 107                         0.742
#> 108                         2.542
#> 109                         2.042
#> 110                         2.342
#> 111                         1.342
#> 112                         1.542
#> 113                         1.742
#> 114                         1.242
#> 115                         1.342
#> 116                         1.542
#> 117                         1.742
#> 118                         2.942
#> 119                         3.142
#> 120                         1.242
#> 121                         1.942
#> 122                         1.142
#> 123                         2.942
#> 124                         1.142
#> 125                         1.942
#> 126                         2.242
#> 127                         1.042
#> 128                         1.142
#> 129                         1.842
#> 130                         2.042
#> 131                         2.342
#> 132                         2.642
#> 133                         1.842
#> 134                         1.342
#> 135                         1.842
#> 136                         2.342
#> 137                         1.842
#> 138                         1.742
#> 139                         1.042
#> 140                         1.642
#> 141                         1.842
#> 142                         1.342
#> 143                         1.342
#> 144                         2.142
#> 145                         1.942
#> 146                         1.442
#> 147                         1.242
#> 148                         1.442
#> 149                         1.642
#> 150                         1.342
#> attr(,"assign")
#> [1] 0 1 2 2 2 3 3 3
#> attr(,"contrasts")
#> attr(,"contrasts")$Species
#>            setosa versicolor virginica
#> setosa          1          0         0
#> versicolor      0          1         0
#> virginica       0          0         1
#> 

# lm's standard, baseline-encoded design matrix
model.matrix(lm(Sepal.Length ~ Petal.Length + Species + Petal.Length*Species, data = iris))
#>     (Intercept) Petal.Length Speciesversicolor Speciesvirginica
#> 1             1          1.4                 0                0
#> 2             1          1.4                 0                0
#> 3             1          1.3                 0                0
#> 4             1          1.5                 0                0
#> 5             1          1.4                 0                0
#> 6             1          1.7                 0                0
#> 7             1          1.4                 0                0
#> 8             1          1.5                 0                0
#> 9             1          1.4                 0                0
#> 10            1          1.5                 0                0
#> 11            1          1.5                 0                0
#> 12            1          1.6                 0                0
#> 13            1          1.4                 0                0
#> 14            1          1.1                 0                0
#> 15            1          1.2                 0                0
#> 16            1          1.5                 0                0
#> 17            1          1.3                 0                0
#> 18            1          1.4                 0                0
#> 19            1          1.7                 0                0
#> 20            1          1.5                 0                0
#> 21            1          1.7                 0                0
#> 22            1          1.5                 0                0
#> 23            1          1.0                 0                0
#> 24            1          1.7                 0                0
#> 25            1          1.9                 0                0
#> 26            1          1.6                 0                0
#> 27            1          1.6                 0                0
#> 28            1          1.5                 0                0
#> 29            1          1.4                 0                0
#> 30            1          1.6                 0                0
#> 31            1          1.6                 0                0
#> 32            1          1.5                 0                0
#> 33            1          1.5                 0                0
#> 34            1          1.4                 0                0
#> 35            1          1.5                 0                0
#> 36            1          1.2                 0                0
#> 37            1          1.3                 0                0
#> 38            1          1.4                 0                0
#> 39            1          1.3                 0                0
#> 40            1          1.5                 0                0
#> 41            1          1.3                 0                0
#> 42            1          1.3                 0                0
#> 43            1          1.3                 0                0
#> 44            1          1.6                 0                0
#> 45            1          1.9                 0                0
#> 46            1          1.4                 0                0
#> 47            1          1.6                 0                0
#> 48            1          1.4                 0                0
#> 49            1          1.5                 0                0
#> 50            1          1.4                 0                0
#> 51            1          4.7                 1                0
#> 52            1          4.5                 1                0
#> 53            1          4.9                 1                0
#> 54            1          4.0                 1                0
#> 55            1          4.6                 1                0
#> 56            1          4.5                 1                0
#> 57            1          4.7                 1                0
#> 58            1          3.3                 1                0
#> 59            1          4.6                 1                0
#> 60            1          3.9                 1                0
#> 61            1          3.5                 1                0
#> 62            1          4.2                 1                0
#> 63            1          4.0                 1                0
#> 64            1          4.7                 1                0
#> 65            1          3.6                 1                0
#> 66            1          4.4                 1                0
#> 67            1          4.5                 1                0
#> 68            1          4.1                 1                0
#> 69            1          4.5                 1                0
#> 70            1          3.9                 1                0
#> 71            1          4.8                 1                0
#> 72            1          4.0                 1                0
#> 73            1          4.9                 1                0
#> 74            1          4.7                 1                0
#> 75            1          4.3                 1                0
#> 76            1          4.4                 1                0
#> 77            1          4.8                 1                0
#> 78            1          5.0                 1                0
#> 79            1          4.5                 1                0
#> 80            1          3.5                 1                0
#> 81            1          3.8                 1                0
#> 82            1          3.7                 1                0
#> 83            1          3.9                 1                0
#> 84            1          5.1                 1                0
#> 85            1          4.5                 1                0
#> 86            1          4.5                 1                0
#> 87            1          4.7                 1                0
#> 88            1          4.4                 1                0
#> 89            1          4.1                 1                0
#> 90            1          4.0                 1                0
#> 91            1          4.4                 1                0
#> 92            1          4.6                 1                0
#> 93            1          4.0                 1                0
#> 94            1          3.3                 1                0
#> 95            1          4.2                 1                0
#> 96            1          4.2                 1                0
#> 97            1          4.2                 1                0
#> 98            1          4.3                 1                0
#> 99            1          3.0                 1                0
#> 100           1          4.1                 1                0
#> 101           1          6.0                 0                1
#> 102           1          5.1                 0                1
#> 103           1          5.9                 0                1
#> 104           1          5.6                 0                1
#> 105           1          5.8                 0                1
#> 106           1          6.6                 0                1
#> 107           1          4.5                 0                1
#> 108           1          6.3                 0                1
#> 109           1          5.8                 0                1
#> 110           1          6.1                 0                1
#> 111           1          5.1                 0                1
#> 112           1          5.3                 0                1
#> 113           1          5.5                 0                1
#> 114           1          5.0                 0                1
#> 115           1          5.1                 0                1
#> 116           1          5.3                 0                1
#> 117           1          5.5                 0                1
#> 118           1          6.7                 0                1
#> 119           1          6.9                 0                1
#> 120           1          5.0                 0                1
#> 121           1          5.7                 0                1
#> 122           1          4.9                 0                1
#> 123           1          6.7                 0                1
#> 124           1          4.9                 0                1
#> 125           1          5.7                 0                1
#> 126           1          6.0                 0                1
#> 127           1          4.8                 0                1
#> 128           1          4.9                 0                1
#> 129           1          5.6                 0                1
#> 130           1          5.8                 0                1
#> 131           1          6.1                 0                1
#> 132           1          6.4                 0                1
#> 133           1          5.6                 0                1
#> 134           1          5.1                 0                1
#> 135           1          5.6                 0                1
#> 136           1          6.1                 0                1
#> 137           1          5.6                 0                1
#> 138           1          5.5                 0                1
#> 139           1          4.8                 0                1
#> 140           1          5.4                 0                1
#> 141           1          5.6                 0                1
#> 142           1          5.1                 0                1
#> 143           1          5.1                 0                1
#> 144           1          5.9                 0                1
#> 145           1          5.7                 0                1
#> 146           1          5.2                 0                1
#> 147           1          5.0                 0                1
#> 148           1          5.2                 0                1
#> 149           1          5.4                 0                1
#> 150           1          5.1                 0                1
#>     Petal.Length:Speciesversicolor Petal.Length:Speciesvirginica
#> 1                              0.0                           0.0
#> 2                              0.0                           0.0
#> 3                              0.0                           0.0
#> 4                              0.0                           0.0
#> 5                              0.0                           0.0
#> 6                              0.0                           0.0
#> 7                              0.0                           0.0
#> 8                              0.0                           0.0
#> 9                              0.0                           0.0
#> 10                             0.0                           0.0
#> 11                             0.0                           0.0
#> 12                             0.0                           0.0
#> 13                             0.0                           0.0
#> 14                             0.0                           0.0
#> 15                             0.0                           0.0
#> 16                             0.0                           0.0
#> 17                             0.0                           0.0
#> 18                             0.0                           0.0
#> 19                             0.0                           0.0
#> 20                             0.0                           0.0
#> 21                             0.0                           0.0
#> 22                             0.0                           0.0
#> 23                             0.0                           0.0
#> 24                             0.0                           0.0
#> 25                             0.0                           0.0
#> 26                             0.0                           0.0
#> 27                             0.0                           0.0
#> 28                             0.0                           0.0
#> 29                             0.0                           0.0
#> 30                             0.0                           0.0
#> 31                             0.0                           0.0
#> 32                             0.0                           0.0
#> 33                             0.0                           0.0
#> 34                             0.0                           0.0
#> 35                             0.0                           0.0
#> 36                             0.0                           0.0
#> 37                             0.0                           0.0
#> 38                             0.0                           0.0
#> 39                             0.0                           0.0
#> 40                             0.0                           0.0
#> 41                             0.0                           0.0
#> 42                             0.0                           0.0
#> 43                             0.0                           0.0
#> 44                             0.0                           0.0
#> 45                             0.0                           0.0
#> 46                             0.0                           0.0
#> 47                             0.0                           0.0
#> 48                             0.0                           0.0
#> 49                             0.0                           0.0
#> 50                             0.0                           0.0
#> 51                             4.7                           0.0
#> 52                             4.5                           0.0
#> 53                             4.9                           0.0
#> 54                             4.0                           0.0
#> 55                             4.6                           0.0
#> 56                             4.5                           0.0
#> 57                             4.7                           0.0
#> 58                             3.3                           0.0
#> 59                             4.6                           0.0
#> 60                             3.9                           0.0
#> 61                             3.5                           0.0
#> 62                             4.2                           0.0
#> 63                             4.0                           0.0
#> 64                             4.7                           0.0
#> 65                             3.6                           0.0
#> 66                             4.4                           0.0
#> 67                             4.5                           0.0
#> 68                             4.1                           0.0
#> 69                             4.5                           0.0
#> 70                             3.9                           0.0
#> 71                             4.8                           0.0
#> 72                             4.0                           0.0
#> 73                             4.9                           0.0
#> 74                             4.7                           0.0
#> 75                             4.3                           0.0
#> 76                             4.4                           0.0
#> 77                             4.8                           0.0
#> 78                             5.0                           0.0
#> 79                             4.5                           0.0
#> 80                             3.5                           0.0
#> 81                             3.8                           0.0
#> 82                             3.7                           0.0
#> 83                             3.9                           0.0
#> 84                             5.1                           0.0
#> 85                             4.5                           0.0
#> 86                             4.5                           0.0
#> 87                             4.7                           0.0
#> 88                             4.4                           0.0
#> 89                             4.1                           0.0
#> 90                             4.0                           0.0
#> 91                             4.4                           0.0
#> 92                             4.6                           0.0
#> 93                             4.0                           0.0
#> 94                             3.3                           0.0
#> 95                             4.2                           0.0
#> 96                             4.2                           0.0
#> 97                             4.2                           0.0
#> 98                             4.3                           0.0
#> 99                             3.0                           0.0
#> 100                            4.1                           0.0
#> 101                            0.0                           6.0
#> 102                            0.0                           5.1
#> 103                            0.0                           5.9
#> 104                            0.0                           5.6
#> 105                            0.0                           5.8
#> 106                            0.0                           6.6
#> 107                            0.0                           4.5
#> 108                            0.0                           6.3
#> 109                            0.0                           5.8
#> 110                            0.0                           6.1
#> 111                            0.0                           5.1
#> 112                            0.0                           5.3
#> 113                            0.0                           5.5
#> 114                            0.0                           5.0
#> 115                            0.0                           5.1
#> 116                            0.0                           5.3
#> 117                            0.0                           5.5
#> 118                            0.0                           6.7
#> 119                            0.0                           6.9
#> 120                            0.0                           5.0
#> 121                            0.0                           5.7
#> 122                            0.0                           4.9
#> 123                            0.0                           6.7
#> 124                            0.0                           4.9
#> 125                            0.0                           5.7
#> 126                            0.0                           6.0
#> 127                            0.0                           4.8
#> 128                            0.0                           4.9
#> 129                            0.0                           5.6
#> 130                            0.0                           5.8
#> 131                            0.0                           6.1
#> 132                            0.0                           6.4
#> 133                            0.0                           5.6
#> 134                            0.0                           5.1
#> 135                            0.0                           5.6
#> 136                            0.0                           6.1
#> 137                            0.0                           5.6
#> 138                            0.0                           5.5
#> 139                            0.0                           4.8
#> 140                            0.0                           5.4
#> 141                            0.0                           5.6
#> 142                            0.0                           5.1
#> 143                            0.0                           5.1
#> 144                            0.0                           5.9
#> 145                            0.0                           5.7
#> 146                            0.0                           5.2
#> 147                            0.0                           5.0
#> 148                            0.0                           5.2
#> 149                            0.0                           5.4
#> 150                            0.0                           5.1
#> attr(,"assign")
#> [1] 0 1 2 2 3 3
#> attr(,"contrasts")
#> attr(,"contrasts")$Species
#> [1] "contr.treatment"
#>