lmabc: Linear Regression with Abundance-Based Constraints (ABCs)

lmabc provides estimation and inference for linear regression with categorical covariates (race, sex, etc.). Common strategies, including the defaults in lm, select a "reference" group (e.g,. White, Male, etc.) and present all results relative to this group. However, this approach suffers from alarming biases, statistical inefficiencies, and limited interpretability. lmabc uses abundance-based constraints (ABCs) that (1) fully eliminate these biases, (2) provide appealing statistical properties related to estimation invariance and improved statistical power, and (3) produce more interpretable model output. lmabc reproduces the core functionality of lm and includes functions for generalized (logistic, Poisson, etc.) and penalized (lasso, ridge, etc.) linear models. Details are provided in Kowal (2024) doi:10.1101/2024.01.04.23300033 .

See also

Author

Maintainer: Dan Kowal daniel.r.kowal@gmail.com

Authors:

Prayag Gordy

Virginia Baskin

Jai Uparkar

Other contributors:

Caleb Fikes [contributor]