Fitting Linear Models with Abundance-Based Constraints (ABCs)

lmabc is used to fit linear models using abundance-based constraints (ABCs). ABCs provide more equitable and interpretable output for regression models with categorical covariates.

Usage

lmabc(formula, data, ..., cprobs = NULL)

Arguments

formula

an object of class "formula()" (or one that can be coerced to that class); a symbolic description of the model to be fitted.

data

a data frame (or object coercible by as.data.frame to a data frame) containing the variables in the model.

...

For lm(): additional arguments to be passed to the low level regression fitting functions (see below).

cprobs

an optional named list with an entry for each named categorical variable in the model, specifying the probabilities of each category, which must sum to 1. By default, cprobs will be calculated from the proportions in the data.

Details

An lmabc model is specified identically to the corresponding lm model. At this time, lmabc only supports a single response variable, and the data must be passed into the data parameter.

Differences from lm

The default approach for linear regression with categorical covariates is reference group encoding (RGE): one category is selected as the "reference group" for each categorical variable, and all results are presented relative to this group. This output is inequitable and can be misleading, especially for categorical covariates like race/ethnicity, gender identity, religion, national origin, etc.

For example, suppose an analyst fits a model of the form y ~ x + race + x:race, where x is a continuous covariate. This model allows for race-specific coefficients on \(x\). However, RGE requires a reference group, typically White (or selected alphabetically), and the lm output is presented relative to this group: (Intercept) refers to the intercept for the White group and x refers to the coefficients on \(x\) for the White group. The remaining race-specific parameters are all presented as differences relative to the White group. Clearly, this output is inequitable: it elevates one group (White) above others, and presents this group as "normal" while others are "deviations from normal". The output is also unclear: that the (Intercept) and x effects refer only to the reference (White) group is nowhere mentioned in any default output.

lmabc addresses these issues. ABCs parametrize the regression model so that the main effect terms, here (Intercept) and x, are averages of the race-specific terms. For instance, the coefficients on \(x\) represent the race-averaged effect of \(x\), and the race-specific coefficients on \(x\) represent deviations from the average, rather than from the reference (White) group. The notion of "average" derives from the argument cprob: these can be input by the user (e.g., population proportions), otherwise they will be set to the sample proportions for each group.

Value

lmabc returns an object of class "lmabc." Many generics commonly used for lm objects have been implemented for lmabc: summary, coefficients, plot, predict, and more. See the DESCRIPTION file for all implemented S3 methods.

See also

Examples

fit <- lmabc(Sepal.Length ~ Petal.Length + Species + Petal.Length*Species, data = iris)
summary(fit)
#> 
#> Call:
#> lmabc(formula = Sepal.Length ~ Petal.Length + Species + Petal.Length * 
#>     Species, data = iris)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -0.73479 -0.22785 -0.03132  0.24375  0.93608 
#> 
#> Coefficients:
#>                                 Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)                     5.524317   0.220516  25.052  < 2e-16 ***
#> Petal.Length                    0.788771   0.102549   7.692 2.09e-12 ***
#> Speciessetosa                   0.726787   0.428933   1.694  0.09235 .  
#> Speciesversicolor              -0.004114   0.224189  -0.018  0.98538    
#> Speciesvirginica               -0.722672   0.239830  -3.013  0.00305 ** 
#> Petal.Length:Speciessetosa     -0.246478   0.189867  -1.298  0.19631    
#> Petal.Length:Speciesversicolor  0.039510   0.118337   0.334  0.73896    
#> Petal.Length:Speciesvirginica   0.206968   0.114212   1.812  0.07205 .  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.3365 on 144 degrees of freedom
#> Multiple R-squared:  0.8405,	Adjusted R-squared:  0.8349 
#> F-statistic: 151.7 on 5 and 144 DF,  p-value: < 2.2e-16
#> 

predict(fit, newdata = data.frame(Petal.Length = 1.5, Species = "setosa"))
#>        1 
#> 5.026607